+++ app/models/data_source.rb
@@ -1,9 +1,7 @@
 class DataSource < ApplicationRecord
   include ModelAttributes::Name
+  include ServiceProvided
   include DataSource::Type
-
-  belongs_to :configured_service
-  validates :configured_service, presence: true
 
   has_many :data_source_alerts
   has_many :alerts, through: :data_source_alerts
